Move tests for get_default_primary_search_objective into automl utils tests
