+++ evalml/tests/automl_tests/test_automl.py
@@ -10,11 +10,7 @@ import woodwork as ww
 from sklearn.model_selection import KFold, StratifiedKFold
 
 from evalml import AutoMLSearch
-from evalml.automl import (
-    TimeSeriesSplit,
-    TrainingValidationSplit,
-    get_default_primary_search_objective
-)
+from evalml.automl import TimeSeriesSplit, TrainingValidationSplit
 from evalml.automl.callbacks import (
     log_and_save_error_callback,
     log_error_callback,
@@ -31,13 +27,7 @@ from evalml.data_checks import (
 from evalml.demos import load_breast_cancer, load_wine
 from evalml.exceptions import AutoMLSearchException, PipelineNotFoundError
 from evalml.model_family import ModelFamily
-from evalml.objectives import (
-    R2,
-    CostBenefitMatrix,
-    FraudCost,
-    LogLossBinary,
-    LogLossMulticlass
-)
+from evalml.objectives import CostBenefitMatrix, FraudCost
 from evalml.objectives.utils import get_core_objectives, get_objective
 from evalml.pipelines import (
     BinaryClassificationPipeline,
@@ -1518,17 +1508,6 @@ def test_data_split_multi(X_y_multi):
 
     y[5] = 0
     automl.search(X, y, data_checks="disabled")
-
-
-def test_get_default_primary_search_objective():
-    assert isinstance(get_default_primary_search_objective("binary"), LogLossBinary)
-    assert isinstance(get_default_primary_search_objective(ProblemTypes.BINARY), LogLossBinary)
-    assert isinstance(get_default_primary_search_objective("multiclass"), LogLossMulticlass)
-    assert isinstance(get_default_primary_search_objective(ProblemTypes.MULTICLASS), LogLossMulticlass)
-    assert isinstance(get_default_primary_search_objective("regression"), R2)
-    assert isinstance(get_default_primary_search_objective(ProblemTypes.REGRESSION), R2)
-    with pytest.raises(KeyError, match="Problem type 'auto' does not exist"):
-        get_default_primary_search_objective("auto")
 
 
 @patch('evalml.tuners.skopt_tuner.SKOptTuner.add')
